+++ Shared/View/Company/CompaniesListView.swift
@@ -1,0 +1,31 @@
+//
+//  CompaniesListView.swift
+//  Story
+//
+//  Created by Alexandre Madeira on 07/05/23.
+//
+
+import SwiftUI
+
+struct CompaniesListView: View {
+    let companies: [ProductionCompany]
+    var body: some View {
+        Form {
+            if companies.isEmpty {
+                CenterHorizontalView { ProgressView().padding() }
+            } else {
+                Section {
+                    List(companies, id: \.self) { item in
+                        NavigationLink(value: item) {
+                            Text(item.name)
+                        }
+                    }
+                }
+            }
+        }
+        .navigationTitle("companiesTitle")
+#if os(macOS)
+        .formStyle(.grouped)
+#endif
+    }
+}
