+++ Shared/View/Company/CompanyDetails.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import SDWebImageSwiftUI
 
 struct CompanyDetails: View {
     let company: ProductionCompany
@@ -38,23 +37,6 @@ struct CompanyDetails: View {
                     ShareLink(item: url)
                 }
 #endif
-            }
-            .navigationDestination(for: ItemContent.self) { item in
-                ItemContentDetails(title: item.itemTitle, id: item.id, type: item.itemContentMedia, handleToolbar: true)
-            }
-            .navigationDestination(for: Person.self) { person in
-                PersonDetailsView(title: person.name, id: person.id)
-            }
-            .navigationDestination(for: [String:[ItemContent]].self) { item in
-                let keys = item.map { (key, _) in key }
-                let value = item.map { (_, value) in value }
-                ItemContentCollectionDetails(title: keys[0], items: value[0])
-            }
-            .navigationDestination(for: [Person].self) { items in
-                DetailedPeopleList(items: items)
-            }
-            .navigationDestination(for: ProductionCompany.self) { item in
-                CompanyDetails(company: item)
             }
             ConfirmationDialogView(showConfirmation: $showConfirmation, message: "addedToWatchlist")
         }
@@ -106,7 +88,7 @@ struct CompanyDetails: View {
                 }
             }
         }
-        .padding(.all, settings.isCompactUI ? 10 : nil)
+                  .padding(.all, settings.isCompactUI ? 10 : nil)
     }
 }
 
@@ -115,67 +97,6 @@ struct CompanyDetails_Previews: PreviewProvider {
                                                    id: 125281, logoPath: nil, originCountry: nil, description: nil)
     static var previews: some View {
         CompanyDetails(company: company)
-    }
-}
-
-struct CompaniesListView: View {
-    let companies: [ProductionCompany]
-    var body: some View {
-        Form {
-            if companies.isEmpty {
-                CenterHorizontalView { ProgressView().padding() }
-            } else {
-                Section {
-                    List(companies, id: \.self) { item in
-                        NavigationLink(value: item) {
-                            Text(item.name)
-                        }
-                    }
-                }
-            }
-        }
-        .navigationTitle("companiesTitle")
-#if os(macOS)
-        .formStyle(.grouped)
-#endif
-    }
-}
-
-class CompanyDetailsViewModel: ObservableObject {
-    var page = 1
-    @Published var items = [ItemContent]()
-    @Published var startPagination = false
-    @Published var endPagination = false
-    @Published var isLoaded = false
-    
-    @MainActor
-    func load(_ id: Int) async {
-        do {
-            let movies = try await NetworkService.shared.fetchCompanyFilmography(type: .movie,
-                                                                                 page: page,
-                                                                                 company: id)
-            let shows = try await NetworkService.shared.fetchCompanyFilmography(type: .tvShow,
-                                                                                page: page,
-                                                                                company: id)
-            let result = movies + shows
-            if result.isEmpty {
-                endPagination = true
-                return
-            } else {
-                page += 1
-            }
-            items.append(contentsOf: result.sorted { $0.itemPopularity > $1.itemPopularity })
-            if !startPagination { startPagination = true }
-            if !isLoaded {
-                DispatchQueue.main.async {
-                    self.isLoaded = true
-                }
-            }
-        } catch {
-            if Task.isCancelled { return }
-            let message = "Company ID: \(id), error: \(error.localizedDescription)"
-            CronicaTelemetry.shared.handleMessage(message, for: "CompanyDetailsViewModel.load()")
-        }
     }
 }
